Started over, Ported the SDL2 calls to SDL3 alot more carefully... I reverted the SDL3 port and started from scratch, reading the SDL docs more closely to make sure we actually have feature parity. The difference here is that we actually can change what display the game renders on. Which the previous SDL3 port couldnt do
Whats next:

HighDPI handling, completely rewrite how resolution is handled in SystemIV
Allow stretched resolutions on fullscreen and fullscreen windowed

+++ contrib/systemIV/lib/gucci/input_sdl.cpp
@@ -433,7 +433,7 @@ int InputManagerSDL::EventHandler( unsigned int message, long long wParam, int l
 
 		case SDL_EVENT_KEY_UP:
 		{
-			int keyCode = ConvertSDLKeyIdToWin32KeyId( sdlEvent->key.key );
+			int keyCode = ConvertSDLKeyIdToWin32KeyId( (int)sdlEvent->key.key );
 			AppDebugAssert( keyCode >= 0 && keyCode < KEY_MAX );
 			if ( g_keys[keyCode] != 0 )
 			{
@@ -446,7 +446,7 @@ int InputManagerSDL::EventHandler( unsigned int message, long long wParam, int l
 
 		case SDL_EVENT_KEY_DOWN:
 		{
-			int keyCode = ConvertSDLKeyIdToWin32KeyId( sdlEvent->key.key );
+			int keyCode = ConvertSDLKeyIdToWin32KeyId( (int)sdlEvent->key.key );
 			AppDebugAssert( keyCode >= 0 && keyCode < KEY_MAX );
 			if ( g_keys[keyCode] != 1 )
 			{
